requirements.txt
yfinance
pandas
matplotlib
google-cloud-datastore
google-cloud-storage
